feat: Add nightly build workflow
This commit adds a new workflow file, `.github/workflows/nightly-build.yml`, which sets up a nightly build for the project. The workflow is triggered on pushes to the `development` branch and can also be manually triggered using the workflow dispatch event.

The workflow runs on the latest version of Ubuntu and performs the following steps:
- Checks out the repository at the `development` branch
- Sets up JDK 11 using the Temurin distribution
- Builds the project using Maven
- Renames the generated JAR files with a suffix based on the short commit hash
- Archives the build artifacts
- Notifies success or failure of the build

This workflow will help automate the process of generating nightly builds for the project.