refactor: Update JUnitTestFileGenerator to use relative path for TEST_REPORT_FILE
This commit updates the JUnitTestFileGenerator class to use a relative path for the TEST_REPORT_FILE constant. Instead of directly referencing the exportPath, the code now uses the System.getProperty("user.dir") to get the current working directory and constructs the relative path to the testReport.csv file. This change ensures that the TEST_REPORT_FILE is always correctly resolved, regardless of the location of the project.